bundle/direct: drop MigrateMode, consolidate state-building in bundle/migrate
MigrateMode was a bool parameter on Apply that forked between "save state
without deploying" and "actually deploy". The two MigrateMode(true) callers
(bundle deployment migrate and upload_state_for_yaml_sync) now use
migrate.BuildStateFromTF directly, reading from the local TF state file
without any API calls.

Changes:
- New bundle/migrate/build_state.go: public BuildStateFromTF extracted from
  cmd/bundle/migrate.go, taking *config.Root so callers can pass an
  un-interpolated config (needed by upload_state_for_yaml_sync).
- bundle/direct/bundle_apply.go: drop MigrateMode type and parameter; Apply
  now only handles real deployments.
- bundle/phases/{deploy,destroy}.go: drop MigrateMode(false) argument.
- upload_state_for_yaml_sync.go: replace CalculatePlan+Apply with
  BuildStateFromTF; keep reverseInterpolate since config is TF-interpolated.
- cmd/bundle/deployment/migrate.go: replace CalculatePlan+Apply with
  BuildStateFromTF; drop exported RunPlanCheck/GetCommonArgs wrappers.
- cmd/bundle/migrate.go: delegate to migrate.BuildStateFromTF.
